backend/requirements.txt
acme==5.3.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certbot==5.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
ConfigArgParse==1.7.1
configobj==5.0.9
cryptography==46.0.5
Deprecated==1.3.1
distro==1.9.0
ecdsa==0.19.1
fastapi==0.133.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
iniconfig==2.3.0
josepy==2.2.0
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.2
opencv-python-headless==4.13.0.92
packaging==26.0
parsedatetime==2.6
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
PyPDF2==3.0.1
pypdfium2==5.5.0
pyRFC3339==2.1.0
pytesseract==0.3.13
pytest==9.0.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
reportlab==4.4.10
requests==2.32.5
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.47
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wrapt==2.1.1
